Termino obtener lecciones por curso mencionando los estudiantes que tienen acceso a dichas lecciones.
