+++ routes/lessons.js
@@ -50,4 +50,16 @@ router.delete('/delete/:uuid', async (req, res) => {
     }
 });
 
+router.get('/lessons-by-course/:uuid', async (req, res)=> {
+    const keyAuthentication = await apiKey.listApi(req, res);
+    if (keyAuthentication) {
+        lessonsController.getLessonsByCourse(req, res);
+    } else {
+        return res.status(400).json({
+            ok: false,
+            mensaje: 'Error de Autenticación ApiKey',
+        });
+    }
+});
+
 module.exports = router;
